requirements.txt
streamlit==1.31.1
pandas==2.2.0
pillow==10.2.0
streamlit-lottie==0.0.5
requests==2.31.0
apify-client==1.9.*
markdown2==2.5.*
pdfkit==1.0.0
wkhtmltopdf==0.2

langchain==0.3.*
langchain-core==0.3.*
langchain-community==0.3.*
langchain_google_genai==2.1.*
langchain_openai==0.3.*
langchain-text-splitters==0.3.*

langsmith==0.1.*
pydantic==2.9.*
python-dotenv==1.0.*
uvicorn==0.29.*
weasyprint==65.*
supabase==2.15.*
python-decouple==3.*
sendgrid==6.11.*
xhtml2pdf
gnews==0.4.*
sentry-sdk==1.35.0
